Luna/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# luna.py: 5,6,7
Pillow == 5.1.0

# luna.py: 3
beautifulsoup4 == 4.6.0

# luna.py: 4
prettytable == 0.7.2

# luna.py: 1
requests == 2.18.4
